+++ packages/graphql-armor/src/envelop/protections/max-aliases.ts
@@ -2,7 +2,9 @@ import { maxAliasesPlugin } from '@escape.tech/graphql-armor-max-aliases';
 
 import { EnvelopConfigurationEnhancement, EnvelopProtection } from './base-protection';
 
-export class EnvelopMaxAliasesProtection extends EnvelopProtection {
+export class EnvelopMaxAliasesProtection<
+  PluginContext extends Record<string, any> = {},
+> extends EnvelopProtection<PluginContext> {
   get isEnabled(): boolean {
     if (!this.config.maxAliases) {
       return this.enabledByDefault;
@@ -10,7 +12,7 @@ export class EnvelopMaxAliasesProtection extends EnvelopProtection {
     return this.config.maxAliases.enabled ?? this.enabledByDefault;
   }
 
-  protect(): EnvelopConfigurationEnhancement {
+  protect(): EnvelopConfigurationEnhancement<PluginContext> {
     return {
       plugins: [maxAliasesPlugin(this.config.maxAliases)],
     };