+++ packages/plugins/max-aliases/src/index.ts
@@ -20,7 +20,7 @@ type MaxAliasesOptions = {
 
 const maxAliasesDefaultOptions: Required<MaxAliasesOptions> = {
   n: 15,
-  allowList: [],
+  allowList: ['__typename'],
   exposeLimits: true,
   errorMessage: 'Query validation error.',
   onAccept: [],
@@ -78,8 +78,8 @@ class MaxAliasesVisitor {
     if (
       'alias' in node &&
       node.alias &&
-      node.name.value !== '__typename' &&
-      !this.config.allowList.includes(node.alias.value)
+      !this.config.allowList.includes(node.alias.value) &&
+      !this.config.allowList.includes(node.name.value)
     ) {
       ++aliases;
     }
@@ -109,7 +109,9 @@ class MaxAliasesVisitor {
 const maxAliasesRule = (options?: MaxAliasesOptions) => (context: ValidationContext) =>
   new MaxAliasesVisitor(context, options);
 
-const maxAliasesPlugin = (options?: MaxAliasesOptions): Plugin => {
+const maxAliasesPlugin = <PluginContext extends Record<string, any> = {}>(
+  options?: MaxAliasesOptions,
+): Plugin<PluginContext> => {
   return {
     onValidate({ addValidationRule }: any) {
       addValidationRule(maxAliasesRule(options));