+++ packages/plugins/max-aliases/test/index.spec.ts
@@ -1,9 +1,14 @@
+import type { Plugin } from '@envelop/types';
 import { assertSingleExecutionValue, createTestkit } from '@envelop/testing';
 import { makeExecutableSchema } from '@graphql-tools/schema';
 import { describe, expect, it } from '@jest/globals';
 import { jest } from '@jest/globals';
 
 import { maxAliasesPlugin } from '../src/index';
+
+// test checking if the plugin inherits the context correctly
+const _test_0: Plugin = maxAliasesPlugin();
+const _test_1: Plugin<{ my: 'ctx' }> = maxAliasesPlugin();
 
 const typeDefinitions = `
   type Book {
@@ -101,6 +106,26 @@ describe('maxAliasesPlugin', () => {
     expect(result.data).toEqual({
       getBook: null,
     });
+  });
+
+  it('counts __typename aliases against limit when not included in allowList', async () => {
+    const maxAliases = 1;
+    const allowList = [];
+    const testkit = createTestkit([maxAliasesPlugin({ n: maxAliases, allowList })], schema);
+    const result = await testkit.execute(/* GraphQL */ `
+      query A {
+        getBook(title: "null") {
+          typenameA: __typename
+          typenameB: __typename
+        }
+      }
+    `);
+
+    assertSingleExecutionValue(result);
+    expect(result.errors).toBeDefined();
+    expect(result.errors?.map((error) => error.message)).toEqual([
+      `Syntax Error: Aliases limit of ${maxAliases} exceeded, found ${maxAliases + 1}.`,
+    ]);
   });
 
   it('respects fragment aliases', async () => {